requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.3.0
langchain-community>=0.3.0
tavily-python>=0.5.0
stripe>=10.0.0
googlemaps>=4.10.0
openpyxl>=3.1.0
pydantic>=2.0.0
python-dotenv>=1.0.0
rich>=13.0.0
requests>=2.31.0
